Converted functions to methods of class in order to be able to inherit class and override methods.
