+++ mail_archives/views/templates.xml
@@ -6,6 +6,7 @@
                   inherit_id="web.assets_backend">
             <xpath expr="." position="inside">
                 <script src="/mail_archives/static/src/js/archives.js" type="text/javascript"></script>
+                <script src="/mail_archives/static/src/js/base.js" type="text/javascript"></script>
             </xpath>
         </template>
     </data>
